+++ rollup.config.js
@@ -9,12 +9,16 @@ import json from '@rollup/plugin-json';
 import babel from 'rollup-plugin-babel';
 import sourceMaps from 'rollup-plugin-sourcemaps';
 import glslify from 'rollup-plugin-glslify';
+import replace from '@rollup/plugin-replace';
 
 const pkgObj = require('./package.json');
 
 function getExternals(pkg) {
-  const { peerDependencies = {} } = pkg;
-  return Object.keys(peerDependencies);
+  const { devDependencies = {}, dependencies = {} } = pkg;
+  const externals = Object.keys(devDependencies).concat(
+    Object.keys(dependencies)
+  );
+  return externals;
 }
 
 export default {
@@ -33,22 +37,24 @@ export default {
     json(),
     babel({
       exclude: 'node_modules/**',
-      presets: ['@babel/env', '@babel/preset-react']
+      presets: ['@babel/env', '@babel/preset-react'],
+      runtimeHelpers: true
     }),
     glslify({ basedir: 'src/layers/XRLayer' }),
     // Allow bundling cjs modules (unlike webpack, rollup doesn't understand cjs)
-    commonjs({
-      namedExports: {
-        // left-hand side can be an absolute path, a path
-        // relative to the current directory, or the name
-        // of a module in node_modules
-        'node_modules/geotiff/dist/geotiff.bundle.min.js': ['fromUrl', 'Pool']
-      }
-    }),
+    commonjs(),
     // Allow node_modules resolution, so you can use 'external' to control
     // which external modules to include in the bundle
     // https://github.com/rollup/rollup-plugin-node-resolve#usage
     resolve(),
+    replace({
+      delimiters: ['', ''],
+      values: {
+        "require('readable-stream/transform')": "require('stream').Transform",
+        'require("readable-stream/transform")': 'require("stream").Transform',
+        'readable-stream': 'stream'
+      }
+    }),
     // Resolve source maps to the original source
     sourceMaps()
   ]