+++ src/layers/index.js
@@ -1,3 +1,4 @@
 export { default as VivViewerLayer } from './VivViewerLayer';
 export { default as StaticImageLayer } from './StaticImageLayer';
 export { default as OverviewLayer } from './OverviewLayer';
+export { default as ScaleBarLayer } from './ScaleBarLayer';