+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- OMETiffLoader for reading ome-tiff files directly
+
 ### Changed
 
 ## 0.1.6
@@ -16,6 +18,7 @@
 ### Changed
 
 - Refactored VivViewer to take in arbitrary views/layers
+- Check for `loader.getRaster` in `VivViewerLayer` before rendering `StaticImageLayer` base
 
 ## 0.1.5